+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Implement `list` command to list available machines from machines.yml
 * Display syntax when `build` called with no machine
+* Support running multiple tasks - e.g. `machines build machine passenger passenger_nginx nginx`
 
 
 == 0.5.2